pyproject.toml
[project]
name = "visitaid"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.2.1",
    "vision-agents[deepgram,elevenlabs,gemini,getstream]>=0.3.8",
]
